Change structure of symptomList to symptomMap so is easier and faster to perform the action on the app.
